Django_Project/AMARSHOP/amarshop/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
django-filter